ignores the swagger validation test
it's not worth figuring out the necessary CI setup to get go onto jenkins
to actually run this test. It would also require figuring out the world
setup for GOPATH and whatnot. On top of that, the swagger validation is run
in the go build using the file generated from this tool. So this test is now
only good for fast turn-around running locally.