Remove generated SDKs
Ruby and PHP SDKs are now public and open-source. We should treat the
public Github repos as the canonical SDKs, as opposed to these.

We also may want to think about how we will want to keep private and
public versions of api.json in sync with each other, but we could
probably worry about later if we find the need to make several public
revisions to the public APIs.