Parameterize templates to accept Swagger title
Also drops in two caviar-specific documents for the
API conventions and the changelog. I am not sure if this
is the right way to do this, so very open to suggestions
and thoughts on this.

We put up a version of the docs and I'm pretty happy with it. It is not
as pretty or well thought out as Barlow's docs.connect.squareup.com, but
it gets the job done!