Fix ProtoOptions.getValue() functions when it is matched against option names with additional scopes.
The original intent of the `endsWith` check in the getValue() functions should be matching the specified `optionName` against the potentially scoped `option.name()`. However, the order was reversed so it was not able to match any scoped names.

The bug is found when parsing the `OAuth` protos, where some options are in the form of `squareup.connect.v2.common.message_status` rather than `common.message_status`.